refactoring code
